+++ exif/__init__.py
@@ -1,3 +1,4 @@
 """Read and modify image EXIF metadata using Python."""
 
+from exif._constants import ColorSpace, Saturation, Sharpness
 from exif._image import Image