+++ exif/_hex_interface.py
@@ -1,6 +1,6 @@
 """Hexadecimal string interface module."""
 
-from exif._constants import HEX_PER_BYTE
+from exif._constants import EXIF_BIG_ENDIAN_HEADER, EXIF_LITTLE_ENDIAN_HEADER, HEX_PER_BYTE
 
 
 class HexInterface(object):
@@ -8,7 +8,13 @@ class HexInterface(object):
     """Hexadecimal string interface class."""
 
     def __init__(self, hex_string):
+        self.endianness = EXIF_BIG_ENDIAN_HEADER
         self._hex_string = hex_string
+
+    @staticmethod
+    def _change_endian(hexadecimal_string):
+        return "".join(reversed(
+            [hexadecimal_string[i:i+2] for i in range(0, len(hexadecimal_string), 2)]))
 
     def delete(self, offset, num_bytes):
         """Delete a sequence of hexadecimal digits from the hexadecimal string.
@@ -47,11 +53,13 @@ class HexInterface(object):
     def modify_hex(self, offset, hex_string):
         """Replace bytes in the hexadecimal string with a new string.
 
-
         :param int offset: starting byte to modify
-        :param int hex_string: bytes to overwrite with
+        :param str hex_string: bytes to overwrite with
 
         """
+        if self.endianness == EXIF_LITTLE_ENDIAN_HEADER:
+            hex_string = self._change_endian(hex_string)
+
         self._hex_string = (
             self._hex_string[:offset * HEX_PER_BYTE] +
             hex_string +
@@ -59,16 +67,21 @@ class HexInterface(object):
         )
 
     def modify_number(self, offset, num_bytes, value):
-        """Replace byets in the hexadecimal string with a number.
+        """Replace bytes in the hexadecimal string with a number.
 
         :param int offset: starting byte to modify
         :param int num_bytes: number of bytes to store value in
         :param int value: new numeric value
 
         """
+        number_hex_bytes = hex(value).lstrip('0x').zfill(num_bytes * HEX_PER_BYTE)
+
+        if self.endianness == EXIF_LITTLE_ENDIAN_HEADER:
+            number_hex_bytes = self._change_endian(number_hex_bytes)
+
         self._hex_string = (
             self._hex_string[:offset * HEX_PER_BYTE] +
-            hex(value).lstrip('0x').zfill(num_bytes * HEX_PER_BYTE) +
+            number_hex_bytes +
             self._hex_string[(offset + num_bytes) * HEX_PER_BYTE:]
         )
 
@@ -79,7 +92,25 @@ class HexInterface(object):
         :param int num_bytes: number of bytes to read
 
         """
-        return self._hex_string[offset * HEX_PER_BYTE:(offset + num_bytes) * HEX_PER_BYTE]
+        hex_bytes = self._hex_string[offset * HEX_PER_BYTE:(offset + num_bytes) * HEX_PER_BYTE]
+
+        if self.endianness == EXIF_LITTLE_ENDIAN_HEADER:
+            hex_bytes = self._change_endian(hex_bytes)
+
+        if len(hex_bytes) != num_bytes * HEX_PER_BYTE:
+            raise ValueError("byte offset exceeds length of segment")
+
+        return hex_bytes
+
+    def set_endianness(self, endian_constant):
+        """Set the endianness of the byte string.
+
+        :param endian_constant: endian marker constant
+        :type endian_constant: str (EXIF_BIG_ENDIAN_HEADER or EXIF_LITTLE_ENDIAN_HEADER)
+
+        """
+        assert endian_constant in [EXIF_BIG_ENDIAN_HEADER, EXIF_LITTLE_ENDIAN_HEADER]
+        self.endianness = endian_constant
 
     def wipe(self, offset, num_bytes):
         """Wipe clear bytes in the hexadecimal string.