+++ docs/release_notes.rst
@@ -2,9 +2,25 @@
 Release Notes
 #############
 
-********************************************
-[0.3.0] Add dunder dir support. (2018-12-26)
-********************************************
+***********************************************
+[0.3.1] Fix little endian support. (2018-02-10)
+***********************************************
+
+Previously, this package did not fully support little endian EXIF metadata in images, raising ``ValueError`` exceptions.
+Now, reading EXIF hexadecimal strings and values takes endianness into account.
+
+This release includes the following under-the-hood changes:
+
+- Move tag reading and modification functions into the IFD tag class.
+- Add enumerations for color space, sharpness, and saturation as a proof-of-concept for leveraging enumerations. (More
+  enumerations coming soon in a future release!)
+- Improve test coverage.
+
+See https://github.com/TNThieding/exif/issues/5 for more information.
+
+*******************************************
+[0.3.0] Add ``dir()`` support. (2018-12-26)
+*******************************************
 
 Implement mechanism for listing EXIF tags in an image using ``dir()``.